backend/pyproject.toml
[project]
name = "ai-grader-backend"
version = "0.1.0"
description = "AI Homework Photo Feedback Checker - Backend"
requires-python = ">=3.11"

dependencies = [
    "fastapi==0.104.1",
    "uvicorn==0.24.0",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    "sqlalchemy==2.0.23",
    "psycopg[binary]==3.19.1",
    "python-multipart==0.0.6",
    "pillow==10.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "black==23.12.0",
    "ruff==0.1.8",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
